% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{fusion_operations}
\alias{fusion_operations}
\title{Build the operation registry}
\usage{
fusion_operations(...)
}
\arguments{
\item{...}{Named \code{\link[=fusion_operation]{fusion_operation()}} objects (bare functions are wrapped as aggregate operations).}
}
\value{
An object of class \code{fusion_operations}.
}
\description{
Build the operation registry
}
