% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/destination.R
\name{fusion_peers}
\alias{fusion_peers}
\title{List peers}
\usage{
fusion_peers(fusion)
}
\arguments{
\item{fusion}{A \code{fusion} object.}
}
\value{
A data frame with columns \code{peer}, \code{legId}, \code{direction}, \code{state}.
}
\description{
List peers
}
